bmp -> pgm

+++ ex01/bitmap.h
@@ -4,7 +4,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-void saveImage(char* imageName, float **img, int imageWidth, int imageHeight);
-void saveColoredImage(char* imageName, float **r, float **g, float **b, int imageWidth, int imageHeight);
+void saveImage(char* imageName, int **img, int imageWidth, int imageHeight);
+void saveColoredImage(char* imageName, int **r, int **g, int **b, int imageWidth, int imageHeight);
 
 #endif
